+++ pakyow-routing/lib/pakyow/routing/controller.rb
@@ -423,13 +423,11 @@ module Pakyow
           data = file_or_data
 
           if file_or_data.is_a?(File)
-            @connection.set_header("content-length", file_or_data.size)
             type ||= MiniMime.lookup_by_filename(file_or_data.path)&.content_type.to_s
           end
 
           @connection.set_header("content-type", type || DEFAULT_SEND_TYPE)
         elsif file_or_data.is_a?(String)
-          @connection.set_header("content-length", file_or_data.bytesize)
           @connection.set_header("content-type", type) if type
           data = StringIO.new(file_or_data)
         else